+++ server/Procfile
@@ -1,0 +1,1 @@
+web: gunicorn -b :$PORT tabby:app
