+++ server/tabby_server/__init__.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+def create_app_instance(name:str = "Tabby_Server"):
+    return (Flask(name))
