--- server/tabby_server/vision/example_yolo/yolo_outputs_here
@@ -1,2 +1,0 @@
-Blank text file so a folder is made to hold the yolo outputs.
-Eventually we will not need this. But is a shortcut for now.
